ci: publish nightly traces to S3 instead of the ci-data repo
sgl-project/ci-data hit ~100 GB and stopped accepting pushes on Jul 21, so
nightly traces have been silently dropped since then. Traces now go to S3
with a 30-day lifecycle rule, uploaded via a GitHub OIDC role.

Also fixes TRACE_BASE_URL, which still pointed at the retired
sglang-bot/sglang-ci-data repo — every trace link in the nightly report has
404'd since May.

The GitHub-repo upload helpers move to scripts/ci/utils/github_repo_upload.py
because the diffusion publishers still use them.
