+++ scripts/ci/utils/diffusion/publish_comparison_results.py
@@ -20,10 +20,10 @@ import time
 from datetime import datetime, timezone
 from pathlib import Path
 
-# Reuse GitHub API helpers from publish_traces.
+# Reuse GitHub API helpers from github_repo_upload.
 # Support both direct script execution and package-style imports.
 if __package__:
-    from ..publish_traces import (
+    from ..github_repo_upload import (
         create_blobs,
         create_commit,
         create_tree,
@@ -36,7 +36,7 @@ if __package__:
     )
 else:
     sys.path.insert(0, str(Path(__file__).resolve().parents[1]))
-    from publish_traces import (
+    from github_repo_upload import (
         create_blobs,
         create_commit,
         create_tree,
